Clean up MuFEB class, use in MupixFEB

+++ common/libmutrig/include/Tiles_FEB.h
@@ -30,8 +30,8 @@ class TilesFEB : public MutrigFEB{
       //Mapping from ASIC number to FPGA_ID and ASIC_ID
       virtual uint16_t FPGAid_from_ID(int asic);
       virtual uint16_t ASICid_from_ID(int asic);
-      virtual uint8_t nModulesPerFEB(){return 1;}
-      virtual uint8_t nAsicsPerModule(){return 13;}
+      virtual uint16_t nModulesPerFEB(){return 1;}
+      virtual uint16_t nAsicsPerModule(){return 13;}
       //Return typeID for building FEB ID map
       virtual FEBTYPE  GetTypeID(){return FEBTYPE::Tile;}
